requirements.txt
elasticsearch~=7.8.0
PyYAML~=5.3.1